Intermediate commit: Rename files, taking care of TODO
