+++ lib/build-registered-datastore-instance.js
@@ -3,7 +3,7 @@
  */
 
 var _ = require('@sailshq/lodash');
-var validateDatastoreConnection = require('./datastore-methods/validate-datastore-connection');
+var checkAdapterCompatibility = require('./check-adapter-compatibility');
 var leaseConnection = require('./datastore-methods/lease-connection');
 var sendStatement = require('./datastore-methods/send-statement');
 var sendNativeQuery = require('./datastore-methods/send-native-query');
@@ -65,14 +65,8 @@ module.exports = function buildRegisteredDatastoreInstance(datastoreName, normal
     // Validate the adapter API version and that the datastore is able
     // to run the lease connection as far as we can tell.
     try {
-      validateDatastoreConnection({
-        datastoreIdentity: datastoreName,
-        adapterApiVersion: adapterApiVersion,
-        adapter: adapter
-      });
-    } catch (e) {
-      return cb(e);
-    }
+      checkAdapterCompatibility(datastoreName, adapter);
+    } catch (e) { return cb(e); }
 
     var adapterDatastore = adapter.datastores[datastoreName];
 
@@ -105,14 +99,8 @@ module.exports = function buildRegisteredDatastoreInstance(datastoreName, normal
     // Validate the adapter API version and that the datastore is able
     // to run the lease connection as far as we can tell.
     try {
-      validateDatastoreConnection({
-        datastoreIdentity: datastoreName,
-        adapterApiVersion: adapterApiVersion,
-        adapter: adapter
-      });
-    } catch (e) {
-      return cb(e);
-    }
+      checkAdapterCompatibility(datastoreName, adapter);
+    } catch (e) { return cb(e); }
 
     var adapterDatastore = adapter.datastores[datastoreName];
 
@@ -154,14 +142,8 @@ module.exports = function buildRegisteredDatastoreInstance(datastoreName, normal
     // Validate the adapter API version and that the datastore is able
     // to run the lease connection as far as we can tell.
     try {
-      validateDatastoreConnection({
-        datastoreIdentity: datastoreName,
-        adapterApiVersion: adapterApiVersion,
-        adapter: adapter
-      });
-    } catch (e) {
-      return cb(e);
-    }
+      checkAdapterCompatibility(datastoreName, adapter);
+    } catch (e) { return cb(e); }
     // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
     var adapterDatastore = adapter.datastores[datastoreName];
@@ -196,14 +178,8 @@ module.exports = function buildRegisteredDatastoreInstance(datastoreName, normal
     // Validate the adapter API version and that the datastore is able
     // to run the lease connection as far as we can tell.
     try {
-      validateDatastoreConnection({
-        datastoreIdentity: datastoreName,
-        adapterApiVersion: adapterApiVersion,
-        adapter: adapter
-      });
-    } catch (e) {
-      return cb(e);
-    }
+      checkAdapterCompatibility(datastoreName, adapter);
+    } catch (e) { return cb(e); }
 
     var adapterDatastore = adapter.datastores[datastoreName];
