+++ lib/check-adapter-compatibility.js
